refactor: renamed news-highlight-card to news-highlights
