+++ src/components/landing/index.ts
@@ -3,7 +3,7 @@
 export { default as Footer } from './footer/footer';
 export { default as Header } from './header/header';
 export { default as Hero } from './hero/hero';
-export { default as NewsHighlightCard } from './news-highlight-card/news-highlight-card';
+export { default as NewsHighlights } from './news-highlights/news-highlights';
 export { default as LandingPage } from './page/landing-page';
 export { default as Quote } from './quote/quote';
 export { default as ScrollShrinkWrapper } from './scroll-shrink-wrapper/scroll-shrink-wrapper';
